# lib/api-zod/src/generated/types/createMarkerInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sonoma County Map API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateMarkerInputCategory } from "./createMarkerInputCategory";

export interface CreateMarkerInput {
  name: string;
  note: string;
  category: CreateMarkerInputCategory;
  lat: number;
  lng: number;
}