# lib/api-zod/src/generated/types/updateMarkerInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sonoma County Map API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateMarkerInputCategory } from "./updateMarkerInputCategory";

export interface UpdateMarkerInput {
  name?: string;
  note?: string;
  category?: UpdateMarkerInputCategory;
}